observability/agentcore-observability-quickstart/requirements.txt
bedrock-agentcore
bedrock-agentcore-starter-toolkit
aws-opentelemetry-distro>=0.10.0
boto3
strands-agents
strands-agents-tools 
strands-agents-builder